Move getLeadTimeWeeks into date field definitions
Only used in these files and colocating the code makes it simpler to
introduce new rules for new covid-19 flow.
