+++ controllers/apply/under10k/fields/project-dates-next.js
@@ -6,9 +6,20 @@ const { oneLine } = require('common-tags');
 const Joi = require('../../lib/joi-extensions');
 const { DateField, RadioField } = require('../../lib/field-types');
 
-const getLeadTimeWeeks = require('../lib/lead-time');
+function getLeadTimeWeeks(country) {
+    const countryLeadTimes = {
+        'england': 18,
+        'northern-ireland': 12,
+        'scotland': 12,
+        'wales': 12,
+    };
+
+    return countryLeadTimes[country] || 18;
+}
 
 module.exports = {
+    _getLeadTimeWeeks: getLeadTimeWeeks, // Exported for tests
+
     fieldProjectStartDateCheck(locale, data = {}) {
         const localise = get(locale);
